# backend/app/api/swagger/spec.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Timeline API",
        "version": "1.0.0",
        "description": "Comprehensive REST API for Timeline project management application. Manage workspaces, projects, tasks, team members, and collaboration features.",
        "contact": {
            "name": "Timeline API Support",
            "email": "api-support@timeline.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "termsOfService": "https://timeline.iamrakesh.codes/terms"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "x-meta": {
                "pathsCount": 46,
                "endpointOperations": 54,
                "models": 8,
                "services": 14,
                "countingRules": "Excluded: /swagger, /swagger-ui, / (root redirect). Included AI endpoints and OAuth endpoints. Counted distinct path templates as pathsCount; each HTTP method on a path counted in endpointOperations."
            },
            "description": "Development server"
        },
        {
            "url": "https://not-known.com/v1",
            "description": "Production server"
        },
        {
            "url": "https://staging-api.timeline.com/v1",
            "description": "Staging server"
        }
    ],
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "JWT Bearer token authentication. Format: 'Bearer {token}'"
            },
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "__session",
                "description": "Session-based authentication using HTTP-only cookies"
            },
            "clerkAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Clerk.js authentication token"
            },
            "apiKeyAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "X-API-Key",
                "description": "API key for server-to-server authentication"
            }
        },
...
    },
    "paths": {
...
        "/admin/users": {
            "get": {
                "summary": "Admin list users (elevated)",
                "tags": [
                    "Admin",
                    "Users"
                ],
                "responses": {
                    "200": {
                        "description": "Users returned"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    }
}